Add color parsing for dxf files, remove unrecognized error
1)
Only 7 colors are supported, assignment is based on the ISO values
from dxfwrite

DXF colors have to be positive integers, negative numbers and 0 are
errors.

Only the colors used in a document will be reported back to the
frontend.

DXF entry, color
1,  red
2,  yellow
3,  green
4,  cyan
5,  blue
6,  magenta
7,  white/black

2)
DXF no longer bails if it sees an unrecognized element.   It
complains, but continues to parse the DXF file.  This lets us skip
over things we might not be able to cut or etch instead of having to
mangle the DXF file until it works.