Flip DXF input on X axis by default
In Testing:

DXF uses 0,0 as the bottom left, but lasersaurs (and some other x, y
tables) use 0,0 as the top right. This brings in DXF files "face down"
-- upside down -- and we end up lasercutting things from the rear and
not the front.   For media with a coated/front side, this means
flipping things over or other rearrangements.

We flip on the x axis then readjust the document if it's moved in to
negative space.

Also removed the error case of parsing the input DXF twice to
convert \r\n to \n