Add support for DXF units in Import
Similar to setting DPI for SVG imports, can now choose between in and
mm for DXF import.  If INSUNITS is set to 0 in a document then it's
up for the person importing the file to set the units in the document.
Metric files from Japan with unit of 0 and US inch files with a unit
of 0 don't play well together.